LayoutTests/inspector/css/css-property-expected.txt

Testing methods of CSSProperty.


== Running test suite: CSSProperty
-- Running test case: CSSProperty.NameCount
PASS: "display" should have at least 2 counts.
PASS: "background-repeat" should have at least 1 count.
PASS: "background-repeat-x" should not be counted.
PASS: "background-repeat-y" should not be counted.
PASS: "background-repeat-invalid" should not be counted.
PASS: "--foo" should not be counted.
PASS: "--bar" should not be counted.

Sorted by usage:
[
  "Fake Property 200 Seen",
  "Fake Property 150 Seen",
  "Fake Property 101 Seen",
  "Fake Property 100 Seen",
  "Fake Property 99 Unseen",
  "Fake Property 99 Seen",
  "Fake Property 1 Unseen",
  "Fake Property 1 Seen",
  "Fake Property 100 Unseen",
  "Fake Property 0 Unseen",
  "Fake Property 0 Seen",
  "Fake Property 101 Unseen",
  "Fake Property 150 Unseen",
  "Fake Property 50 Unseen",
  "Fake Property 50 Seen",
  "Fake Property 200 Unseen"
]
Initially selected completion when ranked by `count % 2`: "Fake Property 101 Seen"
Initially selected completion when ranked by `count % 3`: "Fake Property 200 Seen"

-- Running test case: CSSProperty.prototype.get valid
PASS: "background-repeat" is a valid property.
PASS: "background-repeat-x" is an invalid property.
PASS: "background-repeat-invalid" is an invalid property.
PASS: "--foo" is a valid property.
PASS: "--bar" is a valid property.

-- Running test case: CSSProperty.prototype.get anonymous
PASS: "background-repeat" is not an anonymous CSS property.
PASS: "background-repeat-x" is not an anonymous CSS property.
PASS: "background-repeat-invalid" is not an anonymous CSS property.

-- Running test case: CSSProperty.prototype.get implicit
PASS: "background-repeat" is not an implicit CSS property.
PASS: "background-repeat-x" is not an implicit CSS property.
PASS: "background-repeat-invalid" is not an implicit CSS property.

-- Running test case: CSSProperty.prototype.get value
PASS: "background-repeat" has the value "repeat".
PASS: "background-repeat-x" has the value "repeat".
PASS: "background-repeat-invalid" has the value "repeat".

-- Running test case: CSSProperty.prototype.get enabled
PASS: "background-repeat" is enabled.
PASS: "background-repeat-x" is enabled.
PASS: "background-repeat-invalid" is enabled.
PASS: "background-color" is disabled.

-- Running test case: CSSProperty.prototype.get attached
PASS: "background-repeat" is attached.
PASS: "background-repeat-x" is attached.
PASS: "background-repeat-invalid" is attached.
PASS: "background-color" is detached.

-- Running test case: CSSProperty.prototype.get text
PASS: "background-repeat" has the text "background-repeat: repeat;".
PASS: "background-repeat" has the _text (private) "background-repeat: repeat;".
PASS: "background-repeat-x" has the text "background-repeat-x: repeat;".
PASS: "background-repeat-x" has the _text (private) "background-repeat-x: repeat;".
PASS: "background-repeat-invalid" has the text "background-repeat-invalid: repeat;".
PASS: "background-repeat-invalid" has the _text (private) "background-repeat-invalid: repeat;".

-- Running test case: CSSProperty.prototype.remove
PASS: The removed property should no longer be in properties array.
PASS: The second property should shift and become the first.
